upd tts server

+++ server.py
@@ -4,6 +4,7 @@ import torch
 from fastapi import FastAPI
 from fastapi.responses import FileResponse, Response
 from datetime import datetime
+import time
 
 device = torch.device('cpu')
 torch.set_num_threads(4)
@@ -19,7 +20,7 @@ model = torch.package.PackageImporter(local_file).load_pickle("tts_models", "mod
 model.to(device)
 
 sample_rate = 24000
-speaker='baya'
+speaker=os.getenv("SPEAKER")
 
 temp_audio = 'temp'
 
@@ -29,16 +30,23 @@ if not os.path.exists(temp_audio):
 def clean_temp():
     for filename in os.listdir(temp_audio):
         file_path = os.path.join(temp_audio, filename)
+        last_modified = os.stat(file_path).st_mtime
         if os.path.isfile(file_path) or os.path.islink(file_path):
-            os.unlink(file_path)           
+            if time.time() - last_modified > 10 * 60:
+                os.unlink(file_path)           
 
 
 app = FastAPI()
 
+@app.get("/clean")
+async def clean():
+    clean_temp()
+    return Response({"status": "ok"}, 200)
+
+
 @app.get("/tts_to_wav")
 async def tts(text: str = "", speaker: str = speaker, sample_rate: int = sample_rate):
     try:
-        clean_temp()
         if text == "":
             raise Exception("text is empty")
